Add customisation support for attribute macros
